[meta] Move ApplyTarget/bind to cdsl/inst;
